Speed up FLUX training over 2x with Gaudi optimized attention
